# swagger.json
{"openapi":"3.0.0","paths":{"/api/notes":{"get":{"operationId":"AppController_all","parameters":[{"name":"page","required":true,"in":"query","description":"The page number to retrieve. Must be a positive integer.","schema":{"default":1,"type":"number"}},{"name":"limit","required":true,"in":"query","description":"The limit of data. Must be a positive integer between 1 and 100.","schema":{"default":1,"type":"number"}}],"responses":{"200":{"description":"Successfully retrieved notes."}},"tags":["notes"]},"post":{"operationId":"AppController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateNoteDto"}}}},"responses":{"201":{"description":"Successfully created note."}},"tags":["notes"]}},"/api/notes/{id}":{"get":{"operationId":"AppController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"Successfully retrived note by id."}},"tags":["notes"]},"patch":{"operationId":"AppController_update","parameters":[{"name":"id","required":true,"in":"query","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateNoteDto"}}}},"responses":{"200":{"description":"Successfully updated note."}},"tags":["notes"]},"delete":{"operationId":"AppController_delete","parameters":[{"name":"id","required":true,"in":"query","schema":{"type":"number"}}],"responses":{"204":{"description":"Successfully deleted note."}},"tags":["notes"]}}},"info":{"title":"Note App - API Documentation","description":"This is the API documentation for the Note App, which provides endpoints for managing notes.","version":"1.0.0","contact":{},"license":{"name":"Apache 2.0","url":"http://www.apache.org/licenses/LICENSE-2.0"}},"tags":[],"servers":[],"components":{"schemas":{"CreateNoteDto":{"type":"object","properties":{"title":{"type":"string","example":"drink water!","description":"The title of the note. Must be a string between 5 and 50 characters."},"body":{"type":"string","example":"drink water every 2 hours!","description":"The body of the note. Must be a string between 5 and 500 characters."}},"required":["title","body"]},"UpdateNoteDto":{"type":"object","properties":{}}}}}